+++ source/DSCResources/DSC_DFSNamespaceFolder/DSC_DFSNamespaceFolder.psm1
@@ -18,9 +18,6 @@ $script:localizedData = Get-LocalizedData -DefaultUICulture 'en-US'
 
     .PARAMETER Ensure
     Specifies if the DFS Namespace root should exist.
-
-    .PARAMETER TargetState
-    Specifies the state of the DFS namespace folder target.
 #>
 function Get-TargetResource
 {
@@ -39,12 +36,7 @@ function Get-TargetResource
         [Parameter()]
         [ValidateSet('Present','Absent')]
         [System.String]
-        $Ensure = 'Present',
-
-        [Parameter()]
-        [ValidateSet('Offline','Online')]
-        [System.String]
-        $TargetState = 'Online'
+        $Ensure = 'Present'
     )
 
     Write-Verbose -Message ( @(
@@ -159,6 +151,9 @@ function Get-TargetResource
 
     .PARAMETER ReferralPriorityRank
     Specifies the priority rank, as an integer, for a root target of the DFS namespace.
+
+    .PARAMETER State
+    Specifies the state of the DFS namespace folder.
 #>
 function Set-TargetResource
 {
@@ -181,7 +176,7 @@ function Set-TargetResource
         [Parameter()]
         [ValidateSet('Offline','Online')]
         [System.String]
-        $TargetState = 'Online',
+        $TargetState,
 
         [Parameter()]
         [System.String]
@@ -206,7 +201,12 @@ function Set-TargetResource
 
         [Parameter()]
         [System.UInt32]
-        $ReferralPriorityRank
+        $ReferralPriorityRank,
+
+        [Parameter()]
+        [ValidateSet('Offline','Online')]
+        [System.String]
+        $State
     )
 
     Write-Verbose -Message ( @(
@@ -228,8 +228,16 @@ function Set-TargetResource
             [System.Boolean] $folderChange = $false
 
             # The Folder properties that will be updated
-            $folderProperties = @{
-                State = 'Online'
+            $folderProperties = @{}
+
+            # Check the target properties
+            if (($State) `
+                -and ($folder.State -ne $State))
+            {
+                $folderProperties += @{
+                    State = $State
+                }
+                $folderChange = $true
             }
 
             if (($Description) `
@@ -445,6 +453,9 @@ function Set-TargetResource
 
     .PARAMETER ReferralPriorityRank
     Specifies the priority rank, as an integer, for a root target of the DFS namespace.
+
+    .PARAMETER State
+    Specifies the state of the DFS namespace folder.
 #>
 function Test-TargetResource
 {
@@ -468,7 +479,7 @@ function Test-TargetResource
         [Parameter()]
         [ValidateSet('Offline','Online')]
         [System.String]
-        $TargetState = 'Online',
+        $TargetState,
 
         [Parameter()]
         [System.String]
@@ -493,7 +504,12 @@ function Test-TargetResource
 
         [Parameter()]
         [System.UInt32]
-        $ReferralPriorityRank
+        $ReferralPriorityRank,
+
+        [Parameter()]
+        [ValidateSet('Offline','Online')]
+        [System.String]
+        $State
     )
 
     Write-Verbose -Message ( @(
@@ -517,6 +533,17 @@ function Test-TargetResource
             # The Namespace Folder exists and should
 
             # Check the Namespace parameters
+            if (($State) `
+                -and ($folder.State -ne $State))
+            {
+                Write-Verbose -Message ( @(
+                    "$($MyInvocation.MyCommand): "
+                    $($script:localizedData.NamespaceFolderParameterNeedsUpdateMessage) `
+                            -f $Path,'State'
+                    ) -join '' )
+                $desiredConfigurationMatch = $false
+            }
+
             if (($Description) `
                 -and ($folder.Description -ne $Description))
             {