+++ source/Examples/Resources/DFSNamespaceRoot/1-DFSNamespaceRoot_Domain_MultipleTarget_Config.ps1
@@ -56,6 +56,7 @@ Configuration DFSNamespaceRoot_Domain_MultipleTarget_Config
         DFSNamespaceRoot DFSNamespaceRoot_Domain_Software_CA
         {
             Path                 = '\\contoso.com\software'
+            State                = 'Online'
             TargetPath           = '\\ca-fileserver\software'
             Ensure               = 'Present'
             Type                 = 'DomainV2'
@@ -66,6 +67,7 @@ Configuration DFSNamespaceRoot_Domain_MultipleTarget_Config
         DFSNamespaceRoot DFSNamespaceRoot_Domain_Software_MA
         {
             Path                 = '\\contoso.com\software'
+            State                = 'Online'
             TargetPath           = '\\ma-fileserver\software'
             Ensure               = 'Present'
             Type                 = 'DomainV2'
@@ -76,6 +78,7 @@ Configuration DFSNamespaceRoot_Domain_MultipleTarget_Config
         DFSNamespaceRoot DFSNamespaceRoot_Domain_Software_NY
         {
             Path                 = '\\contoso.com\software'
+            State                = 'Online'
             TargetPath           = '\\ny-fileserver\software'
             Ensure               = 'Present'
             Type                 = 'DomainV2'
@@ -87,6 +90,7 @@ Configuration DFSNamespaceRoot_Domain_MultipleTarget_Config
         DFSNamespaceFolder DFSNamespaceFolder_Domain_SoftwareIT_CA
         {
             Path                 = '\\contoso.com\software\it'
+            State                = 'Online'
             TargetPath           = '\\ca-fileserver\it'
             Ensure               = 'Present'
             Description          = 'AD Domain based DFS namespace for storing IT specific software installers'
@@ -96,6 +100,7 @@ Configuration DFSNamespaceRoot_Domain_MultipleTarget_Config
         DFSNamespaceFolder DFSNamespaceFolder_Domain_SoftwareIT_MA
         {
             Path                 = '\\contoso.com\software\it'
+            State                = 'Online'
             TargetPath           = '\\ma-fileserver\it'
             Ensure               = 'Present'
             Description          = 'AD Domain based DFS namespace for storing IT specific software installers'
@@ -105,6 +110,7 @@ Configuration DFSNamespaceRoot_Domain_MultipleTarget_Config
         DFSNamespaceFolder DFSNamespaceFolder_Domain_SoftwareIT_NY_01
         {
             Path                 = '\\contoso.com\software\it'
+            State                = 'Online'
             TargetPath           = '\\ny-fileserver01\it'
             Ensure               = 'Present'
             Description          = 'AD Domain based DFS namespace for storing IT specific software installers'
@@ -114,6 +120,7 @@ Configuration DFSNamespaceRoot_Domain_MultipleTarget_Config
         DFSNamespaceFolder DFSNamespaceFolder_Domain_SoftwareIT_NY02
         {
             Path                 = '\\contoso.com\software\it'
+            State                = 'Online'
             TargetPath           = '\\ny-fileserver02\it'
             Ensure               = 'Present'
             TargetState          = 'Offline'