+++ tests/Integration/DSC_DFSNamespaceFolder.config.ps1
@@ -12,6 +12,7 @@ Configuration DSC_DFSNamespaceFolder_Config {
             EnableTargetFailback         = $Node.EnableTargetFailback
             ReferralPriorityClass        = $Node.ReferralPriorityClass
             ReferralPriorityRank         = $Node.ReferralPriorityRank
+            State                        = $Node.State
         }
     }
 }